Binutils scripts: Remove --disable-multilib.
While Binutils shares the same build system as GCC, Binutils doesn't
actually use the multilib options at all. While the flag or lack thereof
doesn't actually cause a build failure, it is a useless parameter that
doesnt't need to get passed. It's better to just have it removed.

GCC does make use of the flag passed to that build system and does
really matter, which is why this PR doesn't mess with that. That param
should remain.

Big thanks for Xi Ruoyao from the Linux From Scratch project making it
known. It has been decently far out. Decided to check the scripts of
this project again and noticed the flags are still here, so this PR
cleans things up a little bit to only have what we need.