MinGW-w64 Scripts: Only build for the target.
This commit/PR forces the build system of MinGW-w64 to disable building
for an architecture that is not the target. While this commit forces the
build system to not do so on i686-w64-mingw32, it isn't strictly
necessary, but may be needed in the future.

As for x86_64-w32-mingw32, what happens is for the C runtime section of
the package is it builds for lib and lib32. Since we don't want
multilib, the existence of the lib32 libraries is useless and takes up
about 120M in the extracted tarball hosted on the Linux From Scratch
website. Future proofing has been done for Winpthreads, as while the
flag is not needed for Winpthreads, it can end up in the future building
lib32 static libraries.

All in all, this commit will shave off 126M from 1579M, resulting in a
new total size being 1453M, and with the future proofing should keep it
at around that size, MinGW-w64 updates permitting.

For a future note, unrelated to this commit, we should also explain the
--with-default-msvcrt=msvcrt parameter does in the build scripts. I can
make a future PR pretty soon that does so, as well as making sure the
options appear in the right order.