scripts: Add variables.in. Fixes #5.
This makes updating any packages in the toolchain less of a burden by
offloading all the package versions as variables into one single file
that gets sourced by each script. This in turn allows for a single
prefix for the opt directory that is shared between each architecture,
which then will differ further in each script, but will no longer
contain any package versions. At that point, it's arch-specific.

An important note is that while the scripts for the most part won't have
to get updated whenever a package is updated in the toolchain besides
the actual variables.in file, README.md must be updated everytime.
Markdown doesn't support variables, so we will have to offload the
version-specific instructions to another/more script(s) if we want to
get around that.

Tested all scripts besides 015-nsis.sh due to a lack of scons.
Everything else results in a working toolchain for both architectures.